[utils] move adjacent pairs to utils
